Problem Set 8 - Ellipsodal Calculator
Your goal for this problem set is to create a useful spreadsheet in the same spirit as the GPA Calculator.  You are free to choose the topic but will need to get it approved by your teacher first.  The spreadsheet should allow the user to enter in information which the spreadsheet then processes and manipulates to produce useful output.  The spreadsheet should have the following features:

Instructions must be presented in the spreadsheet cells on how to use the spreadsheet (what should be entered, where it should be entered, what will be computed, etc.).
The spreadsheet must be aesthetically formatted with appropriate font sizes, background colors, borders, etc.  Make the spreadsheet look inviting to use.  Appropriate use of colors and borders help guide the user when he or she enters information.
No raw error message should be displayed in cells when unexpected or erroneous cell content is entered. For example, if you’re getting a #DIV/0! error message, format it appropriately using IF conditionals so a clean or instructive response is displayed instead to the user.
Your spreadsheet should make relevant use of IF conditionals, mathematical operations (sum, average, etc.), cell formatting rules (background colors, font color, etc.), and at least one lookup table.  Feel free to use other commands as you see fit.